requirements.txt
streamlit==1.39.0
pillow==10.4.0
ImageHash==4.3.1
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.1
matplotlib==3.8.4
openpyxl==3.1.5
plotly==5.24.1
streamlit-aggrid==0.3.5
reportlab==4.2.5
qrcode==7.4.2
rapidfuzz==3.9.6
streamlit-aggrid==0.3.5
openpyxl==3.1.5

